pyproject.toml
[project]
name = "meeting-intelligence"
version = "0.1.0"
description = "Real-time meeting transcription and analysis platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.4",
    "alembic>=1.13.1",

    # Task queue
    "celery[redis]>=5.3.6",
    "redis>=5.0.1",

    # ML/AI
    "openai-whisper>=20231117",
    "torch>=2.1.0",
    "transformers>=4.37.0",
    "sentence-transformers>=2.3.1",

    # Audio processing
    "pydub>=0.25.1",
    "soundfile>=0.12.1",

    # Utils
    "pydantic[email]>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "certifi>=2024.0.0",

    # LLM APIs
    "google-genai>=1.0.0",

    # Cloud storage
    "boto3>=1.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.13",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
diarization = [
    "pyannote.audio>=3.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "workers"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
